add deleteCookies function to navBar.tsx
